Feature/creating model entities

+++ RadmsDataModels/Models/AccidentDetailsTransaction.cs
@@ -15,8 +15,8 @@ namespace RadmsDataModels.Models
             VictimDetailsTransactions = new HashSet<VictimDetailsTransaction>();
         }
 
-        public int Lat { get; set; }
-        public int Long { get; set; }
+        public double Lat { get; set; }
+        public double Long { get; set; }
         [Key]
         [Column("AccidentID", TypeName = "numeric(18, 0)")]
         public decimal AccidentId { get; set; }
